finished

+++ Run.sh
@@ -1,13 +1,4 @@
 #!/bin/bash
-pro=$1
-if [  "$pro" = "" ];then
-	echo "Problem Name(hit enter to use problem1)?";
-	read pro
-	if [  "$pro" = "" ];then
-		pro=problem1
-	fi
-fi
-
 ####
 Err_Ok=0
 Err_Compile=1
@@ -31,6 +22,15 @@ cur=`pwd`
 run=$cur/run
 res=0
 
+pro=$1
+if [  "$pro" = "" ];then
+	echo "Problem Name(hit enter to use problem1)?";
+	read pro
+	if [  "$pro" = "" ];then
+		pro=problem1
+	fi
+fi
+
 if [ ! -d ./problems/$pro ];then
 	echo "Problems(./problems/$pro) DO NOT exist";
 	exit 1
@@ -39,9 +39,14 @@ if [ ! -d ./hws/$pro ];then
 	echo "User files(./hws/$pro) DO NOT exist";
 	exit 1
 fi
-
+doRunAllHws(){
+	echo "No.	Status	Name	Problem	Date";
 for i in $(ls ./hws/$pro/); do
+	if [ ! -d ./hws/$pro/$i ];then
+		continue;
+	fi
 	#echo Checking User: $i
+	name='UnknowName'
 	rm -rf $run/*
 	cp -r ./problems/$pro/test $run/
 	cp ./scripts/test.sh $run/
@@ -54,10 +59,22 @@ for i in $(ls ./hws/$pro/); do
 	else
 		res=10
 	fi
-	###begin run ###
+	if [ -f "./hws/$pro/$i/name.txt" ];then
+		name=`cat ./hws/$pro/$i/name.txt|head -n 1`
+		if [ -z $name ];then
+			name='UnknowName'
+		fi
+	else
+		name='UnknowName'
+	fi
+
+	time=`date +%Y-%m-%d_%H:%M`
 	echo -n $i"	"
 	printErr $res
+	echo -n "	$name	$pro	$time"
 	echo ;
 done
+}
+doRunAllHws |tee ./hws/$pro/result.txt
 
 exit 0
